Move addParameter() to OperationN
